Remove color & move formatNumber to separate file
